+++ src/test/java/com/infosupport/machinelearning/modelmanagement/storage/FileUploadIntegrationTest.java
@@ -2,7 +2,7 @@ package com.infosupport.machinelearning.modelmanagement.storage;
 
 
 import com.infosupport.machinelearning.modelmanagement.controllers.ModelsController;
-import com.infosupport.machinelearning.modelmanagement.storage.repository.ModelRepository;
+import com.infosupport.machinelearning.modelmanagement.repositories.ModelRepository;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
