+++ src/test/java/com/infosupport/machinelearning/modelmanagement/storage/ModelControllerTest.java
@@ -1,7 +1,7 @@
 package com.infosupport.machinelearning.modelmanagement.storage;
 
 import com.infosupport.machinelearning.modelmanagement.controllers.ModelsController;
-import com.infosupport.machinelearning.modelmanagement.storage.repository.ModelRepository;
+import com.infosupport.machinelearning.modelmanagement.repositories.ModelRepository;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.InjectMocks;
