+++ src/test/java/com/infosupport/machinelearning/modelmanagement/storage/ModelRepositoryTest.java
@@ -1,14 +1,13 @@
 package com.infosupport.machinelearning.modelmanagement.storage;
 
 import com.infosupport.machinelearning.modelmanagement.models.Model;
-import com.infosupport.machinelearning.modelmanagement.storage.repository.ModelRepository;
+import com.infosupport.machinelearning.modelmanagement.repositories.ModelRepository;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.io.File;
